The beginnings of the XMBCalendar !
The whole idea is to have a "Calendar" like element, that can
be packaged as a Prefab and placed into any world, without
thinking too much about it.

The Calendar would actually show a curated list of small
worlds (less then 2 megabytes at the moment), with sorting
and filters functions.
When selecting a world in the list, a portal would then
open, allowing the user to jump into the world.

The list of curated world would be sampled through a
specific avatar profile picture.

So far the texture encoding and decoding is almost done.
The C# verison is done, but now I'll have to check if
that's U# compatible.

Signed-off-by: Voyage <voyage@miouyouyou.fr>